Adding final code to new flowrate predictor
