+++ flowrate_predictor_demo.py
@@ -1,7 +1,7 @@
 from path_finding.path_finder import PathFinder
 from file_handling.graph_vertex_edge_init import GraphVertexEdgeInit
 from datetime import date, datetime
-from jh_ml_models.flowrate_predictorV2 import FlowratePredictor
+from jh_ml_models.flowrate_predictor import FlowratePredictor
 if __name__ == "__main__":
     current_time = datetime.now()
     current_time = current_time.replace(month=8)  # We dont have access to data from current month faking that we are actually in the 8th month
